Remove namespace from mailers
