--- app/mailers/peoplefinder/token_mailer.rb
@@ -1,11 +1,0 @@
-require 'peoplefinder'
-
-class Peoplefinder::TokenMailer < ActionMailer::Base
-  layout 'peoplefinder/email'
-
-  def new_token_email(token)
-    @token = token
-
-    mail to: @token.user_email
-  end
-end
