+++ app/mailers/token_mailer.rb
@@ -1,0 +1,9 @@
+class TokenMailer < ActionMailer::Base
+  layout 'email'
+
+  def new_token_email(token)
+    @token = token
+
+    mail to: @token.user_email
+  end
+end
